Xp values now save correctly.

+++ src/main/java/lk1905/gielinorcraft/capability/skill/SkillStorage.java
@@ -16,7 +16,7 @@ public class SkillStorage implements Capability.IStorage<ISkills>{
 		CompoundNBT data = new CompoundNBT();
 		
 		for(int i = 0; i < 26; i++) {
-			data.putInt("xp_" + i, (int) instance.getXp(i));
+			data.putDouble("xp_" + i, instance.getXp(i));
 			data.putInt("static_" + i, instance.getStaticLevel(i));
 			data.putInt("dynamic_" + i, instance.getLevel(i));
 		}
@@ -31,7 +31,7 @@ public class SkillStorage implements Capability.IStorage<ISkills>{
 		CompoundNBT data = (CompoundNBT) nbt;
 
 		for(int i = 0; i < 26; i++) {
-			instance.setXp(i, data.getInt("xp_" + i));
+			instance.setXp(i, data.getDouble("xp_" + i));
 			instance.setStaticLevel(i, data.getInt("static_" + i));
 			instance.setLevel(i, data.getInt("dynamic_" + i));
 		}
